Fixed invalid sidebar layout
Upstream started to use BrowserViewLayout::CalculateContentsContainerLayout()
to get layout value of contents container and side panel.
It's best to override that method but it's difficult because it uses
struct that defined in cc file(ContentsContainerLayoutResult).
So, replaced whole LayoutContentsContainerView() logic with ours instead
of using CalculateContentsContainerLayout().

CalculateContentsContainerLayout() is also used to get proper dialog
position when sidebar state is changed. As it considers panel's size,
it's not exactly fit with ours. We have another sidebar UI also. but the
difference is not much big, I think dialog positioning with
CalculateContentsContainerLayout() would work well.