[cr138][WIP] Comment out CSS override that is triggering DCHECK during printing
We're currently triggering a DCHECK when printing. The print preview window
appears but is blank and immediately closes. There are no relevant errors
printed when this occurs.

Max noticed that the issue goes away when we comment out the two `md-select`
selectors in this CSS override file. Upstream has recently updated parts of the
print UI to use Lit, which may have triggered this issue for us. We're going to
temporarily comment out these selectors while we investigate the root cause
and/or seek out a workaround.